+++ src/pages/MainPage.jsx
@@ -1,12 +1,8 @@
 import styled from "styled-components";
 import RightSidebar from "../components/RightSidebar";
 import { useState } from "react";
-import gifIcon from "../assets/icons/gif.svg";
-import photoIcon from "../assets/icons/photo.svg";
-import locationIcon from "../assets/icons/location.svg";
-import smileIcon from "../assets/icons/smile.svg";
-import fireIcon from "../assets/icons/fire.svg";
 import Items from "./components/mainpage/Items";
+import PostComponents from "./components/mainpage/PostComponents";
 
 const Container = styled.div`
   background: white;
@@ -20,20 +16,25 @@ const MainContainer = styled.div`
   flex-direction: column;
   font-family: system-ui, sans-serif;
   width: 600px;
-  height: 100vh; /* 화면 전체 높이 */
+  height: 100vh;
+  overflow-y: auto;
   border-right: 1px solid rgba(0, 0, 0, 0.1);
   border-left: 1px solid rgba(0, 0, 0, 0.1);
+  &::-webkit-scrollbar {
+    display: none;
+  }
 `;
 
 const TabWrapper = styled.div`
   display: flex;
   border-bottom: 1px solid #ddd;
+  position: sticky;
+  top: 0;
+  z-index: 10;
+  background-color: rgba(255, 255, 255, 0.8); /* 반투명 배경 */
+  backdrop-filter: blur(10px); /* 흐림 효과 */
 `;
-const PostWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  border-bottom: 1px solid #ddd;
-`;
+
 const ShowPostWrapper = styled.div`
   display: flex;
   align-items: center;
@@ -76,56 +77,6 @@ const TabText = styled.span`
     transition: background-color 0.2s ease;
   }
 `;
-const PostWrite = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 10px;
-`;
-const IconsWrite = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-const UserImg = styled.img`
-  width: 35px;
-  height: 35px;
-  border-radius: 20px;
-`;
-const WriteInput = styled.input.attrs({
-  placeholder: "What's happening?",
-})`
-  padding: 12px;
-  font-size: 20px;
-  border: none;
-  outline: none;
-`;
-
-const PostBtn = styled.button`
-  border-radius: 20px;
-  background-color: #a7a7a7;
-  color: white;
-  width: 60px;
-  height: 30px;
-  font-size: 14px;
-  font-weight: bold;
-  border: none;
-  margin: 10px;
-  cursor: pointer;
-  &:hover {
-    background-color: black;
-  }
-`;
-
-const IconsImg = styled.img`
-  width: 20px;
-  height: 20px;
-`;
-const IconsWrapper = styled.div`
-  display: flex;
-  gap: 10px;
-  margin: 10px 60px;
-  cursor: pointer;
-`;
-
 function MainPage() {
   const [activeTab, setActiveTab] = useState("forYou");
   return (
@@ -139,27 +90,14 @@ function MainPage() {
             <TabText active={activeTab === "following"}>Following</TabText>
           </Tab>
         </TabWrapper>
-        <PostWrapper>
-          <PostWrite>
-            <UserImg src="https://mblogthumb-phinf.pstatic.net/MjAyNTAxMjJfMTg4/MDAxNzM3NTQ1NzY1MDc0.Pgcv6JXSxrh1KHXVB2c2X5rJ8FHDrsvsQ5-35AWa0asg.urpL4d0Xau_DkMR-UDeFJT09h0whunwWcezTf9wawu0g.JPEG/image.JPEG?type=w800" />
-            <WriteInput />
-          </PostWrite>
-          <IconsWrite>
-            <IconsWrapper>
-              <IconsImg src={photoIcon} />
-              <IconsImg src={gifIcon} />
-              <IconsImg src={fireIcon} />
-              <IconsImg src={smileIcon} />
-              <IconsImg src={locationIcon} />
-            </IconsWrapper>
-            <PostBtn>post</PostBtn>
-          </IconsWrite>
-        </PostWrapper>
+        <PostComponents />
         {activeTab === "forYou" && (
           <>
             <ShowPostWrapper>
               <span>Show 140 posts</span>
             </ShowPostWrapper>
+            <Items />
+            <Items />
             <Items />
           </>
         )}
